+++ lib/apipie/dsl/definition.rb
@@ -1,0 +1,8 @@
+require "apipie/dsl/base"
+require "apipie/dsl/action"
+require "apipie/dsl/common"
+require "apipie/dsl/param"
+require "apipie/dsl/concern"
+
+require "apipie/dsl/resource"
+require "apipie/dsl/controller"
